changed type in model file; working through resulting errors

+++ pkg/awardqueue/awardqueue_test.go
@@ -393,7 +393,7 @@ func (suite *AwardQueueSuite) Test_AssignTSPsToBands() {
 
 	for i := 0; i < tspsToMake; i++ {
 		tsp, _ := testdatagen.MakeTSP(suite.db, "Test Shipper", testdatagen.RandomSCAC())
-		score := mps + i + 1
+		score := float64(mps + i + 1)
 		testdatagen.MakeTSPPerformance(suite.db, tsp, tdl, nil, score, 0, float64(score)*.5, float64(score)*.5)
 	}
